+++ sdk/cosmos/azure-cosmos-spark-fabric-account-data-resolver/CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Release History
 ### 1.0.0-beta.1 (Unreleased)
-This maven-package will never be released - it is used in fabric for entra id authentication. 
 
 #### Features Added
+* Added account data resolver implementation for Azure Cosmos DB Spark Connector. This allows signed in users in Fabric to resolve account information from the Azure Fabric account, enabling seamless integration with Azure Cosmos DB.
+
 #### Breaking Changes
+
 #### Bugs Fixed
+
 #### Other Changes